requirements.txt
streamlit==1.31.0
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.5
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.1
tiktoken==0.5.2
pypdf==4.0.1
